fix: use unsigned ints

+++ software/relay/globals.h
@@ -1,8 +1,8 @@
 #ifndef GLOBALS_H_
 #define GLOBALS_H_
 
-uint8_t isMaintenanceState;
-uint8_t xisStable;
+unsigned int isMaintenanceState;
+unsigned int xisStable;
 
 void initGlobals();
